zone selectors can have more than one char
This is the third (and last) occurence of this error.
Why do we check zone validity at three different places?

Core::Reader - readers first create a tree and then a bundle
to which it belongs. This is for efficiency reasons
(we don't know whether the tree belongs to the last bundle
or to a new new bundle before the tree is created).
So we cannot use `$bundle->add_tree($root)` here.

Core::Bundle::add_tree - this is the standard way how to create
new trees (within existing bundles).

Core::Node::Root::set_zone - we want to allow changing a zone
of an existing tree. If this methods fails because of invalid
zone, the tree zone is not changed, but the tree remains in
the bundle (and the exception can be catched).
So `$root->set_zone($zone)` cannot be used within
`$bundle->add_tree($root)` because there a failure means
that the tree is not added to the bundle.

We could check the validity in a subroutine called from the three
places, but that would be slow (and the checking code is simple).